+++ app/cli/secrets/secrets_cli_dialogs.go
@@ -1,13 +1,42 @@
 package secrets
 
 import (
+	"errors"
 	"github.com/schollz/progressbar/v3"
 	"k8s-management-go/app/actions/secrets_actions"
 	"k8s-management-go/app/cli/dialogs"
 	"k8s-management-go/app/models"
 	"k8s-management-go/app/utils/loggingstate"
+	"strings"
 )
 
+// ask for password
+func AskForSecretsPassword(passwordText string) (password string, err error) {
+	// Validator for password (keep it simple for now)
+	validate := func(input string) error {
+		if len(input) < 4 {
+			return errors.New("Password too short! ")
+		}
+		if strings.Contains(input, " ") {
+			return errors.New("Password should not contain spaces! ")
+		}
+		return nil
+	}
+
+	// ask for password
+	loggingstate.AddInfoEntry("  -> Ask for the password for secret file...")
+	password, err = dialogs.DialogAskForPassword(passwordText, validate)
+	if err != nil {
+		loggingstate.AddErrorEntryAndDetails("  -> Ask for the password for secret files...failed", err.Error())
+
+		return "", err
+	}
+	loggingstate.AddInfoEntry("  -> Ask for the password for secret file...done")
+
+	return password, err
+}
+
+// apply secrets
 func ApplySecrets() (err error) {
 	// ask for namespace
 	loggingstate.AddInfoEntry("  -> Ask for namespace to apply secrets...")
