+++ app/cli/secrets/secrets_cli_workflow.go
@@ -3,9 +3,7 @@ package secrets
 import (
 	"errors"
 	"k8s-management-go/app/actions/secrets_actions"
-	"k8s-management-go/app/cli/dialogs"
 	"k8s-management-go/app/utils/loggingstate"
-	"strings"
 )
 
 // encrypt the secrets file with given password
@@ -38,6 +36,7 @@ func EncryptSecretsFile() (err error) {
 	return err
 }
 
+// Decrypt secrets file
 func DecryptSecretsFile() (err error) {
 	password, err := AskForSecretsPassword("Password for secrets file")
 	if err != nil {
@@ -45,30 +44,4 @@ func DecryptSecretsFile() (err error) {
 	}
 	err = secrets_actions.ActionDecryptSecretsFile(password)
 	return err
-}
-
-// ask for password
-func AskForSecretsPassword(passwordText string) (password string, err error) {
-	// Validator for password (keep it simple for now)
-	validate := func(input string) error {
-		if len(input) < 4 {
-			return errors.New("Password too short! ")
-		}
-		if strings.Contains(input, " ") {
-			return errors.New("Password should not contain spaces! ")
-		}
-		return nil
-	}
-
-	// ask for password
-	loggingstate.AddInfoEntry("  -> Ask for the password for secret file...")
-	password, err = dialogs.DialogAskForPassword(passwordText, validate)
-	if err != nil {
-		loggingstate.AddErrorEntryAndDetails("  -> Ask for the password for secret files...failed", err.Error())
-
-		return "", err
-	}
-	loggingstate.AddInfoEntry("  -> Ask for the password for secret file...done")
-
-	return password, err
 }
